+++ small_text/integrations/pytorch/datasets.py
@@ -5,7 +5,13 @@ from small_text.base import LABEL_UNLABELED
 from small_text.data import DatasetView
 from small_text.data.datasets import check_size
 from small_text.data.exceptions import UnsupportedOperationException
+from small_text.integrations.pytorch.exceptions import PytorchNotFoundError
 from small_text.utils.labels import list_to_csr
+
+try:
+    import torch
+except ImportError:
+    raise PytorchNotFoundError('Could not import pytorch')
 
 
 class PytorchDataset(ABC):
@@ -109,9 +115,9 @@ class PytorchTextClassificationDataset(PytorchDataset):
             and for multi-label datasets to an empty list.
         vocab : torchtext.vocab.vocab
             Vocabulary object.
-        multi_label : bool
+        multi_label : bool, default=False
             Indicates if this is a multi-label dataset.
-        target_labels : list of int or None
+        target_labels : np.ndarray[int] or None, default=None
             This is a list of (integer) labels to be encountered within this dataset.
             This is important to set if your data does not contain some labels,
             e.g. due to dataset splits, where the labels should however be considered by
@@ -125,7 +131,7 @@ class PytorchTextClassificationDataset(PytorchDataset):
         self._target_labels = None
         if target_labels is not None:
             self.track_target_labels = False
-            self.target_labels = np.array(target_labels)
+            self.target_labels = target_labels
         else:
             self.track_target_labels = True
             self._infer_target_labels()
@@ -235,6 +241,20 @@ class PytorchTextClassificationDataset(PytorchDataset):
             raise ValueError('Cannot remove existing labels from target_labels as long as they '
                              'still exists in the data. Create a new dataset instead.')
         self._target_labels = target_labels
+
+    def clone(self):
+        import copy
+
+        if self.is_multi_label:
+            data = [(torch.clone(d[self.INDEX_TEXT]),
+                     d[self.INDEX_LABEL].copy()) for d in self._data]
+        else:
+            data = [(torch.clone(d[self.INDEX_TEXT]),
+                     np.copy(d[self.INDEX_LABEL])) for d in self._data]
+        return PytorchTextClassificationDataset(data,
+                                                copy.deepcopy(self._vocab),
+                                                multi_label=self.multi_label,
+                                                target_labels=np.copy(self._target_labels))
 
     def to(self, other, non_blocking=False, copy=False):
         """Calls `torch.Tensor.to` on all Tensors in `data`.